move test to correct file
