+++ src/datadoc/backend/tests/test_utils.py
@@ -1,3 +1,4 @@
+import datetime
 import os
 import pathlib
 
@@ -6,6 +7,7 @@ from cloudpathlib.local import LocalGSClient
 from cloudpathlib.local import LocalGSPath
 
 from datadoc.backend.src.utils import calculate_percentage
+from datadoc.backend.src.utils import incorrect_date_order
 from datadoc.backend.src.utils import normalize_path
 from datadoc.backend.tests.constants import DATADOC_METADATA_MODULE_UTILS
 from datadoc.backend.tests.constants import TEST_BUCKET_PARQUET_FILEPATH
@@ -38,3 +40,19 @@ def test_normalize_path(
 
 def test_calculate_percentage():
     assert calculate_percentage(1, 3) == 33  # noqa: PLR2004
+
+
+@pytest.mark.parametrize(
+    ("date_from", "date_until", "expected"),
+    [
+        (datetime.date(2024, 1, 1), datetime.date(1960, 1, 1), True),
+        (datetime.date(1980, 1, 1), datetime.date(2000, 6, 5), False),
+        (None, None, False),
+        (datetime.date(2024, 1, 1), None, False),
+        (None, datetime.date(2024, 1, 1), True),
+        (datetime.date(2024, 1, 1), datetime.date(2024, 1, 1), False),
+    ],
+)
+def test_incorrect_date_order(date_from, date_until, expected):
+    result = incorrect_date_order(date_from, date_until)
+    assert result == expected
