+++ src/datadoc/backend/tests/test_validators.py
@@ -16,26 +16,9 @@ from datadoc.backend.src.model_validation import ObligatoryDatasetWarning
 from datadoc.backend.src.model_validation import ObligatoryVariableWarning
 from datadoc.backend.src.utility.constants import OBLIGATORY_METADATA_WARNING
 from datadoc.backend.src.utility.enums import TemporalityTypeType
-from datadoc.backend.src.utils import incorrect_date_order
 
 if TYPE_CHECKING:
     from datadoc.backend.src.core import Datadoc
-
-
-@pytest.mark.parametrize(
-    ("date_from", "date_until", "expected"),
-    [
-        (datetime.date(2024, 1, 1), datetime.date(1960, 1, 1), True),
-        (datetime.date(1980, 1, 1), datetime.date(2000, 6, 5), False),
-        (None, None, False),
-        (datetime.date(2024, 1, 1), None, False),
-        (None, datetime.date(2024, 1, 1), True),
-        (datetime.date(2024, 1, 1), datetime.date(2024, 1, 1), False),
-    ],
-)
-def test_incorrect_date_order(date_from, date_until, expected):
-    result = incorrect_date_order(date_from, date_until)
-    assert result == expected
 
 
 @pytest.mark.parametrize(
@@ -84,7 +67,6 @@ def test_write_metadata_document_created_date(
     assert metadata.dataset.metadata_created_date is not None
 
 
-# TODO(@tilen1976): Frontend?   # noqa: TD003
 @pytest.mark.parametrize(
     ("variable_date", "date_from", "date_until"),
     [
